requirements.txt
anthropic>=0.46
pillow>=11.0
opencv-python>=4.10
numpy>=1.26
moviepy>=1.0
pydub>=0.25
requests>=2.32
imageio>=2.37
imageio-ffmpeg>=0.6
tqdm>=4.67 